Refactor: reorganize network policy imports and remove unused constants
- Consolidated imports from `ldap_protocol.policies.network` for better readability.
- Removed the `constans.py` file as it was no longer needed.
- Updated references to `ProtocolType` and other network policy components across multiple files to reflect the new import structure.
- Ensured all affected files and tests are aligned with the changes in import paths.
