+++ app/ioc.py
@@ -115,8 +115,10 @@ from ldap_protocol.policies.audit.monitor import (
 )
 from ldap_protocol.policies.audit.policies_dao import AuditPoliciesDAO
 from ldap_protocol.policies.audit.service import AuditService
-from ldap_protocol.policies.network.gateway import NetworkPolicyGateway
-from ldap_protocol.policies.network.use_cases import NetworkPolicyUseCase
+from ldap_protocol.policies.network import (
+    NetworkPolicyGateway,
+    NetworkPolicyUseCase,
+)
 from ldap_protocol.policies.password import (
     PasswordPolicyDAO,
     PasswordPolicyUseCases,
