+++ app/ldap_protocol/policies/network/__init__.py
@@ -1,14 +1,22 @@
 """Network policies module."""
 
-from .constans import ProtocolType
-from .dto import NetworkPolicyDTO
-from .exceptions import NetworkPolicyAlreadyExistsError
+from .constants import ProtocolType
+from .dto import NetworkPolicyDTO, NetworkPolicyUpdateDTO, SwapPrioritiesDTO
+from .exceptions import (
+    LastActivePolicyError,
+    NetworkPolicyAlreadyExistsError,
+    NetworkPolicyNotFoundError,
+)
 from .gateway import NetworkPolicyGateway
 from .use_cases import NetworkPolicyUseCase
 
 __all__ = [
     "NetworkPolicyDTO",
+    "NetworkPolicyUpdateDTO",
+    "SwapPrioritiesDTO",
     "NetworkPolicyAlreadyExistsError",
+    "LastActivePolicyError",
+    "NetworkPolicyNotFoundError",
     "NetworkPolicyGateway",
     "NetworkPolicyUseCase",
     "ProtocolType",
